requirements.txt
# Person 1 Requirements — Data Generation & Kafka Producer
# Install with:  pip install -r requirements.txt

# ── Data generation & validation ───────────
numpy>=1.24
pandas>=2.0

# ── Kafka producer & topic management ──────
kafka-python>=2.0.2

# (Person 2's PySpark requirements have been removed from this list)
